Trying to fix the mess that I made with the two previous commits.
Original commit message from CVS:


Trying to fix the mess that I made with the two previous commits.

First commit, doing something wrong (fixing comparisons of signed and unsigned).
Second commit, trying to revert the previous, but changing other things and
reverting unrelated work of other people.
And this third, hopefully fixes it.

+++ plugins/elements/gstfilesink.c
@@ -415,7 +415,7 @@ gst_filesink_chain (GstPad *pad, GstData *_data)
   if (GST_FLAG_IS_SET (filesink, GST_FILESINK_OPEN))
   {
     guint bytes_written = 0, back_pending = 0;
-    if ((guint64) ftell(filesink->file) < filesink->data_written)
+    if (ftell(filesink->file) < filesink->data_written)
       back_pending = filesink->data_written - ftell(filesink->file);
     while (bytes_written < GST_BUFFER_SIZE (buf)) {
       size_t wrote = fwrite (GST_BUFFER_DATA (buf) + bytes_written, 1,
